+++ scripts/setup-local-environment.sh
@@ -1,26 +1,27 @@
 #!/usr/bin/env bash
 set -e
 
-INCLUDE_BASE=${INCLUDE_BASE:-true} # INCLUDE_UTILS must be true to work.
 INCLUDE_UTILS=${INCLUDE_UTILS:-false}
 INCLUDE_DEVELOPER=${INCLUDE_DEVELOPER:-false}
 
 PANOPTES_UTILS=${PANOPTES_UTILS:-$PANDIR/panoptes-utils}
 PANOPTES_POCS=${PANOPTES_POCS:-$PANDIR/POCS}
-_UTILS_IMAGE_URL="gcr.io/panoptes-exp/panoptes-utils:latest"
+UTILS_IMAGE_URL="gcr.io/panoptes-exp/panoptes-utils:latest"
 
-echo "Setting up local environment."
+echo "Setting up local environment in ${PANOPTES_POCS}"
 cd "${PANOPTES_POCS}"
 
 build_utils() {
-  /bin/bash "${PANOPTES_UTILS}/docker/setup-local-environment.sh"
+  INCLUDE_BASE="${INCLUDE_BASE:-true} ""${PANOPTES_UTILS}/scripts/setup-local-environment.sh"
   # Use our local image for build below instead of gcr.io image.
-  _UTILS_IMAGE_URL="panoptes-utils:develop"
+  UTILS_IMAGE_URL="panoptes-utils:develop"
+  echo "Setting UTILS_IMAGE_URL=${UTILS_IMAGE_URL}"
 }
 
 build_develop() {
-  echo "Building local panoptes-pocs:develop from ${_UTILS_IMAGE_URL} in ${PANOPTES_POCS}"
+  echo "Building local panoptes-pocs:develop from ${UTILS_IMAGE_URL} in ${PANOPTES_POCS}"
   docker build \
+    --build-arg "image_url=${UTILS_IMAGE_URL}" \
     -t "panoptes-pocs:develop" \
     -f "${PANOPTES_POCS}/docker/Dockerfile" \
     "${PANOPTES_POCS}"
@@ -29,6 +30,7 @@ build_develop() {
 build_developer() {
   echo "Building local panoptes-pocs:developer from panoptes-pocs:develop in ${PANOPTES_POCS}"
   docker build \
+    --build-arg "userid=$(id -u)" \
     -t "panoptes-pocs:developer" \
     -f "${PANOPTES_POCS}/docker/developer/Dockerfile" \
     "${PANOPTES_POCS}"