+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Changed
 ~~~~~~~
 
 * Add support for taking "dark" frames for cameras with mechanical shutters or opaque filters in the filterwheel. (@AnthonyHorton #989)
+* Updated `panoptes-utils` to `v0.2.27` to support the envvars for starting config server. (@wtgee #1001)
+* Move the `wait-for-it.sh` script into `scripts`. (@wtgee #1001)
 
 
 [0.7.6] - 2020-08-21